proper redirects for slanted floor (#9)
This changes the current test Floor to a slant from (0,0) -> (200,200) and it fixes the `Movement` redirect in the `onCollision` `Adjust` function by checking if the `Mobile`'s x-speed == 0. However there is still a small bug where the `Mobile` falls through the `Floor` unpredictably. I do believe this is caused by a precision error in using Doubles.