Refactors some of the Adjuster logic (#17)
This merges AdjustSupplier up into AdjustHandler, since the supplier wasn't *really* immutable and was basically useless. It also splits the 'adjusters' list into 'incoming' and 'outgoing', which means we have to fix some of the core functions to take an extra parameter. While we're at it, this makes 'adjust()' and 'removeAll() a lot simpler as well.