Solves the dropped messages problem.
In summary, the Game and Room actors would receive an AddPlayer(username) message, figure some shit out, then forward the message and wait for a response. They would then send the response back to the sender. This code looked like this:

(actor ? msg ) foreach { resp => sender ! response }

The problem was that the `sender` variable changes to Deadletters when the ? method returns, so the responses were being sent to the Deadletters actor. There are many ways to solve the problem, but the most elegant (and a much better strategy to begin with) is to simply fire and forget the message with a forward:

actor forward msg

Now when the Player actor receives a Start command, the `sender` will lead directly back to the Application, and we don't have to deal with this response chain up the tree.