Baking a cake with ConnectionServices (#28)
This change attempts to modularize several layers of behavior for a ConnectionService. We now have a PlayConnection that implements the sending and closing behavior. We also have a BufferingConnection and ActorConnection that can be mixed in. It's a good example of the cake pattern at work.

This refactor completely breaks the Application module's WebSocket function. It expects to get an Enumeration back in a Future, and right now that Enumeration is burried in an Actor.